add dynamic routing, fix comments
