+++ src/App.js
@@ -1,14 +1,10 @@
 import React from 'react';
 import { IntlProvider } from 'react-intl';
-import {
-  BrowserRouter, Switch, Redirect, Route,
-} from 'react-router-dom';
-// css
-import 'normalize.css';
-import 'flexboxgrid2/flexboxgrid2.css';
+import { BrowserRouter, Switch, Route } from 'react-router-dom';
 // components
 import Header from './Header';
-import Profile from './Profile';
+import OtherPage from './OtherPage';
+import Profile from './ProfilePage';
 
 export default () => (
   <IntlProvider locale="en">
@@ -16,8 +12,17 @@ export default () => (
       <div className="App">
         <Header />
         <Switch>
-          <Redirect exact from="/" to="/EveryInteract" />
-          <Route exact path="/EveryInteract" component={Profile} />
+          <Route path="/moments" component={OtherPage} />
+          <Route path="/notifications" component={OtherPage} />
+          <Route path="/messages" component={OtherPage} />
+          <Route path="/about" component={OtherPage} />
+          <Route path="/help" component={OtherPage} />
+          <Route path="/terms" component={OtherPage} />
+          <Route path="/privacy" component={OtherPage} />
+          <Route path="/cookies" component={OtherPage} />
+          <Route path="/advertising" component={OtherPage} />
+          <Route path="/info" component={OtherPage} />
+          <Route path="/" component={Profile} />
         </Switch>
       </div>
     </BrowserRouter>
