+++ src/index.js
@@ -1,5 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
+import 'normalize.css';
+import 'flexboxgrid2/flexboxgrid2.css';
 
 ReactDOM.render(<App />, document.getElementById('root'));
